+++ inlang/source-code/sdk/slotfile-webapp/src/main.tsx
@@ -17,6 +17,7 @@ if (isInIframe) {
 	const { MainViewIframe } = await import("./mainViewIframe.js")
 	root.render(
 		<MainViewIframe
+			repoUrl={new URLSearchParams(window.location.search).get("repo")!}
 			projectPath={new URLSearchParams(window.location.search).get("inlangProjectPath")!}
 		/>
 	)