+++ inlang/source-code/sdk/src/v2/rpc-fs/index.ts
@@ -1,22 +1,32 @@
-import { asyncIterableTransferHandler } from "./transfer/asyncIterable.js"
 import * as Comlink from "comlink"
-import type { NodeishFilesystemSubset } from "../types/plugin.js"
+import { asyncIterableTransferHandler } from "./transfer/asyncIterable.js"
 import { watchOptionsTransferHandler } from "./transfer/watchOptions.js"
-import { adapter } from "comlink-node"
+import { nodeishStatsTransferHandler } from "./transfer/nodeishStats.js"
+import type { NodeishFilesystem } from "@lix-js/fs"
 
 Comlink.transferHandlers.set("asyncIterable", asyncIterableTransferHandler)
 Comlink.transferHandlers.set("watchOptions", watchOptionsTransferHandler)
+Comlink.transferHandlers.set("NodeishStats", nodeishStatsTransferHandler)
 
-export function makeFsAvailableTo(fs: NodeishFilesystemSubset, ep: Comlink.Endpoint) {
-	Comlink.expose(fs, adapter(ep))
+export function makeFsAvailableTo(fs: NodeishFilesystem, ep: Comlink.Endpoint) {
+	Comlink.expose(fs, ep)
 }
 
 type FileChangeInfo = { eventType: "rename" | "change"; filename: string | null }
 
-export function getFs(ep: Comlink.Endpoint): NodeishFilesystemSubset {
-	const _fs = Comlink.wrap<NodeishFilesystemSubset>(ep)
+export function getFs(ep: Comlink.Endpoint): NodeishFilesystem {
+	const _fs = Comlink.wrap<NodeishFilesystem>(ep)
 
 	return {
+		_createPlaceholder: _fs._createPlaceholder,
+		_isPlaceholder: _fs._isPlaceholder,
+		readlink: _fs.readlink,
+		stat: _fs.stat,
+		lstat: _fs.lstat,
+		rm: _fs.rm,
+		rmdir: _fs.rmdir,
+		symlink: _fs.symlink,
+		unlink: _fs.unlink,
 		readdir: _fs.readdir,
 		readFile: _fs.readFile as any,
 		writeFile: _fs.writeFile,